centrex-graphql/backend/requirements.txt
strawberry-graphql
uvicorn
sqlalchemy
passlib[bcrypt]
PyJWT
